- Removed all junk. - Removed old files. - Restructured project hierarchy a little bit.
