--- Tripper/Domain/Util/Route Creation/RouteCreationProtocol.swift
@@ -1,14 +1,0 @@
-//
-//  RouteCreationProtocol.swift
-//  Tripper
-//
-//  Created by Denis Cherniy on 12.03.2020.
-//  Copyright © 2020 Denis Cherniy. All rights reserved.
-//
-
-import CoreLocation
-import MapboxDirections
-
-typealias CoordinatesDictionary = Dictionary<String, CLLocationCoordinate2D>
-
-
